Revert "float64"
This reverts commit 7a74445dbb5da5295340e21a9c04ba1f10c29cf3.

+++ scripts/llama_ring_sg/benchmark_ring.py
@@ -45,7 +45,7 @@ def parse_args():
     parser.add_argument("--run_ring_first", action="store_true")
     parser.add_argument("--no-run_ring_first", dest="run_ring_first", action="store_false")
     parser.add_argument("--csv_output_file", type=str, default="benchmark_log.csv", help="Path to the CSV file for logging benchmark results.")
-    parser.add_argument("--dtype", type=str, default="float64", choices=["float64", "float32", "float16", "bfloat16"])
+    parser.add_argument("--dtype", type=str, default="float32", choices=["float32", "float32", "float16", "bfloat16"])
     parser.set_defaults(run_ring_first=True)
     return parser.parse_args()
 
@@ -279,7 +279,7 @@ def main():
         print_memory_snapshot("Initial_Script_Load", initial_mem_snapshot, rank)
     except AttributeError:
         print0(f"[WARNING] Invalid dtype '{args.dtype}', defaulting to float32.")
-        parsed_dtype = torch.float64
+        parsed_dtype = torch.float32
 
     torch.set_default_dtype(parsed_dtype)
